refactor(app): split React Doctor cleanup components
