+++ apps/app/src/react-app/domains/session/modals/model-picker-modal.tsx
@@ -6,45 +6,11 @@ import {
   useMemo,
   useRef,
   useState,
-  type KeyboardEvent as ReactKeyboardEvent,
-  type RefObject,
 } from "react";
-import { CheckCircle2, Circle, Search, X } from "lucide-react";
 
-import { t } from "../../../../i18n";
 import { modelEquals } from "../../../../app/utils";
 import type { ModelOption, ModelRef } from "../../../../app/types";
-
-// Minimal inline provider icon placeholder. The full ProviderIcon gets ported
-// from src/app/components/provider-icon.tsx in a later step of the plan.
-function ProviderIcon({
-  providerId,
-  size = 16,
-  className,
-}: {
-  providerId: string;
-  size?: number;
-  className?: string;
-}) {
-  const initial = providerId.trim().charAt(0).toUpperCase() || "?";
-  return (
-    <span
-      aria-hidden
-      className={className}
-      style={{
-        display: "inline-flex",
-        alignItems: "center",
-        justifyContent: "center",
-        width: size,
-        height: size,
-        fontSize: Math.round(size * 0.65),
-        fontWeight: 600,
-      }}
-    >
-      {initial}
-    </span>
-  );
-}
+import { ModelPickerDialog, type ProviderLinkItem } from "./model-picker-modal-components";
 
 export type ModelPickerModalProps = {
   open: boolean;
@@ -67,13 +33,6 @@ type RenderedItem =
       title: string;
       matchCount: number;
     };
-
-type ProviderLinkItem = {
-  providerID: string;
-  title: string;
-  matchCount: number;
-  index: number;
-};
 
 export function ModelPickerModal(props: ModelPickerModalProps) {
   const searchInputRef = useRef<HTMLInputElement | null>(null);
@@ -302,412 +261,4 @@ export function ModelPickerModal(props: ModelPickerModalProps) {
       onClose={props.onClose}
     />
   );
-}
-
-function ModelPickerDialog(props: {
-  target: "default" | "session";
-  query: string;
-  totalOptions: number;
-  filteredCount: number;
-  current: ModelRef;
-  searchInputRef: RefObject<HTMLInputElement | null>;
-  activeIndex: number;
-  renderedCount: number;
-  recommendedOptions: { opt: ModelOption; index: number }[];
-  otherEnabledOptions: { opt: ModelOption; index: number }[];
-  otherOptions: ProviderLinkItem[];
-  registerOptionRef: (index: number) => (el: HTMLDivElement | null) => void;
-  onSetQuery: (value: string) => void;
-  onSetActiveIndex: (index: number) => void;
-  onSelect: (model: ModelRef) => void;
-  onBehaviorChange: (model: ModelRef, value: string | null) => void;
-  onOpenSettings: () => void;
-  onClose: (options?: { restorePromptFocus?: boolean }) => void;
-}) {
-  return (
-    <div className="fixed inset-0 z-50 bg-gray-1/60 backdrop-blur-sm flex items-start justify-center p-4 overflow-y-auto">
-      <div className="bg-dls-surface border border-dls-border w-full max-w-lg rounded-[24px] shadow-[var(--dls-shell-shadow)] overflow-hidden max-h-[calc(100vh-2rem)] flex flex-col">
-        <div className="p-6 flex flex-col min-h-0">
-          <ModelPickerHeader target={props.target} onClose={props.onClose} />
-          <ModelPickerSearch
-            query={props.query}
-            totalOptions={props.totalOptions}
-            filteredCount={props.filteredCount}
-            searchInputRef={props.searchInputRef}
-            onSetQuery={props.onSetQuery}
-          />
-          <ModelPickerSections
-            current={props.current}
-            activeIndex={props.activeIndex}
-            renderedCount={props.renderedCount}
-            recommendedOptions={props.recommendedOptions}
-            otherEnabledOptions={props.otherEnabledOptions}
-            otherOptions={props.otherOptions}
-            registerOptionRef={props.registerOptionRef}
-            onSetActiveIndex={props.onSetActiveIndex}
-            onSelect={props.onSelect}
-            onBehaviorChange={props.onBehaviorChange}
-            onOpenSettings={props.onOpenSettings}
-            onClose={props.onClose}
-          />
-          <div className="mt-5 flex justify-end shrink-0">
-            <button
-              type="button"
-              className="inline-flex items-center justify-center rounded-full border border-dls-border px-4 py-2 text-[13px] font-medium text-dls-text transition-colors hover:bg-[var(--dls-hover)]"
-              onClick={() => props.onClose()}
-            >
-              {t("settings.done")}
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ModelPickerHeader(props: {
-  target: "default" | "session";
-  onClose: (options?: { restorePromptFocus?: boolean }) => void;
-}) {
-  return (
-    <div className="flex items-start justify-between gap-4">
-      <div>
-        <h3 className="text-lg font-semibold text-gray-12">
-          {t(props.target === "default" ? "model_picker.default_model_title" : "model_picker.chat_model_title")}
-        </h3>
-        <p className="text-sm text-gray-11 mt-1">
-          {t(props.target === "default" ? "model_picker.default_model_desc" : "model_picker.chat_model_desc")}
-        </p>
-      </div>
-      <button
-        type="button"
-        className="inline-flex size-8 items-center justify-center rounded-full text-gray-10 transition-colors hover:bg-[var(--dls-hover)]"
-        onClick={() => props.onClose()}
-      >
-        <X size={16} />
-      </button>
-    </div>
-  );
-}
-
-function ModelPickerSearch(props: {
-  query: string;
-  totalOptions: number;
-  filteredCount: number;
-  searchInputRef: RefObject<HTMLInputElement | null>;
-  onSetQuery: (value: string) => void;
-}) {
-  return (
-    <div className="mt-5">
-      <div className="relative">
-        <Search size={16} className="absolute left-3 top-1/2 -translate-y-1/2 text-dls-secondary" />
-        <input
-          ref={props.searchInputRef}
-          type="text"
-          value={props.query}
-          onChange={(event) => props.onSetQuery(event.currentTarget.value)}
-          placeholder={t("settings.search_models")}
-          className="w-full bg-dls-surface border border-dls-border rounded-xl py-2.5 pl-9 pr-3 text-sm text-dls-text placeholder:text-dls-secondary focus:outline-none focus:ring-1 focus:ring-[rgba(var(--dls-accent-rgb),0.2)] focus:border-dls-accent"
-        />
-      </div>
-      {props.query.trim() ? (
-        <div className="mt-2 text-xs text-dls-secondary">
-          {t("settings.showing_models", { count: props.filteredCount, total: props.totalOptions })}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function ModelPickerSections(props: {
-  current: ModelRef;
-  activeIndex: number;
-  renderedCount: number;
-  recommendedOptions: { opt: ModelOption; index: number }[];
-  otherEnabledOptions: { opt: ModelOption; index: number }[];
-  otherOptions: ProviderLinkItem[];
-  registerOptionRef: (index: number) => (el: HTMLDivElement | null) => void;
-  onSetActiveIndex: (index: number) => void;
-  onSelect: (model: ModelRef) => void;
-  onBehaviorChange: (model: ModelRef, value: string | null) => void;
-  onOpenSettings: () => void;
-  onClose: (options?: { restorePromptFocus?: boolean }) => void;
-}) {
-  return (
-    <div className="mt-4 space-y-4 overflow-y-auto pr-1 -mr-1 min-h-0">
-      <ModelOptionsSection
-        title={t("model_picker.recommended")}
-        options={props.recommendedOptions}
-        current={props.current}
-        activeIndex={props.activeIndex}
-        registerOptionRef={props.registerOptionRef}
-        onSetActiveIndex={props.onSetActiveIndex}
-        onSelect={props.onSelect}
-        onBehaviorChange={props.onBehaviorChange}
-      />
-      <ModelOptionsSection
-        title={t("model_picker.other_connected_models")}
-        options={props.otherEnabledOptions}
-        current={props.current}
-        activeIndex={props.activeIndex}
-        registerOptionRef={props.registerOptionRef}
-        onSetActiveIndex={props.onSetActiveIndex}
-        onSelect={props.onSelect}
-        onBehaviorChange={props.onBehaviorChange}
-      />
-      {props.otherOptions.length > 0 ? (
-        <section className="space-y-2">
-          <div className="px-1 text-[11px] font-semibold uppercase tracking-[0.12em] text-gray-9">
-            {t("model_picker.more_providers")}
-          </div>
-          {props.otherOptions.map((provider) => (
-            <ProviderLinkRow
-              key={provider.providerID}
-              provider={provider}
-              activeIndex={props.activeIndex}
-              registerOptionRef={props.registerOptionRef}
-              onSetActiveIndex={props.onSetActiveIndex}
-              onOpenSettings={props.onOpenSettings}
-              onClose={props.onClose}
-            />
-          ))}
-        </section>
-      ) : null}
-      {props.renderedCount === 0 ? (
-        <div className="rounded-2xl border border-gray-6/70 bg-gray-1/40 px-4 py-6 text-sm text-gray-10">
-          {t("model_picker.no_results")}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function ModelOptionsSection(props: {
-  title: string;
-  options: { opt: ModelOption; index: number }[];
-  current: ModelRef;
-  activeIndex: number;
-  registerOptionRef: (index: number) => (el: HTMLDivElement | null) => void;
-  onSetActiveIndex: (index: number) => void;
-  onSelect: (model: ModelRef) => void;
-  onBehaviorChange: (model: ModelRef, value: string | null) => void;
-}) {
-  if (props.options.length === 0) return null;
-  return (
-    <section className="space-y-2">
-      <div className="px-1 text-[11px] font-semibold uppercase tracking-[0.12em] text-gray-9">
-        {props.title}
-      </div>
-      {props.options.map(({ opt, index }) => (
-        <ModelOptionRow
-          key={`${opt.providerID}/${opt.modelID}`}
-          opt={opt}
-          index={index}
-          activeIndex={props.activeIndex}
-          current={props.current}
-          registerOptionRef={props.registerOptionRef}
-          onSetActiveIndex={props.onSetActiveIndex}
-          onSelect={props.onSelect}
-          onBehaviorChange={props.onBehaviorChange}
-        />
-      ))}
-    </section>
-  );
-}
-
-function ModelOptionRow(props: {
-  opt: ModelOption;
-  index: number;
-  activeIndex: number;
-  current: ModelRef;
-  registerOptionRef: (index: number) => (el: HTMLDivElement | null) => void;
-  onSetActiveIndex: (index: number) => void;
-  onSelect: (model: ModelRef) => void;
-  onBehaviorChange: (model: ModelRef, value: string | null) => void;
-}) {
-  const { opt } = props;
-  const active = modelEquals(props.current, {
-    providerID: opt.providerID,
-    modelID: opt.modelID,
-  });
-  const isKeyboardActive = props.index === props.activeIndex;
-
-  const selectOption = () => props.onSelect({ providerID: opt.providerID, modelID: opt.modelID });
-
-  return (
-    <div
-      role="button"
-      tabIndex={0}
-      ref={props.registerOptionRef(props.index)}
-      className={[
-        "group w-full text-left rounded-xl px-3 py-2.5 transition-colors cursor-pointer",
-        active
-          ? "bg-gray-3 text-gray-12"
-          : isKeyboardActive
-            ? "bg-gray-2 text-gray-12"
-            : "text-gray-10 hover:bg-gray-1/70 hover:text-gray-11",
-      ].join(" ")}
-      onMouseEnter={() => props.onSetActiveIndex(props.index)}
-      onClick={selectOption}
-      onKeyDown={(event: ReactKeyboardEvent<HTMLDivElement>) => {
-        if (event.key !== "Enter" && event.key !== " ") return;
-        if (event.nativeEvent.isComposing) return;
-        event.preventDefault();
-        selectOption();
-      }}
-    >
-      <div className="flex items-start gap-3">
-        <ProviderIcon
-          providerId={opt.providerID}
-          size={16}
-          className={[
-            "mt-[1px] shrink-0 transition-colors",
-            active ? "text-gray-12" : "text-gray-10 group-hover:text-gray-11",
-          ].join(" ")}
-        />
-        <div className="flex-1 min-w-0">
-          <div
-            className={[
-              "text-[13px] flex items-center justify-between gap-2",
-              active ? "font-medium text-gray-12" : "text-current",
-            ].join(" ")}
-          >
-            <span className="truncate">{opt.title}</span>
-          </div>
-          <div
-            className={[
-              "mt-0.5 flex items-center gap-3 text-[11px]",
-              active ? "text-gray-10" : "text-gray-9 group-hover:text-gray-10",
-            ].join(" ")}
-          >
-            <span className="truncate">{opt.description ?? opt.providerID}</span>
-            <span className="ml-auto opacity-70 font-mono">
-              {opt.providerID}/{opt.modelID}
-            </span>
-          </div>
-          {opt.footer ? (
-            <div className={["text-[11px] mt-1", active ? "text-gray-10" : "text-gray-8 group-hover:text-gray-9"].join(" ")}>
-              {opt.footer}
-            </div>
-          ) : null}
-          {active && (opt.behaviorOptions?.length ?? 0) > 0 ? (
-            <ModelBehaviorOptions
-              opt={opt}
-              onBehaviorChange={props.onBehaviorChange}
-            />
-          ) : null}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ModelBehaviorOptions(props: {
-  opt: ModelOption;
-  onBehaviorChange: (model: ModelRef, value: string | null) => void;
-}) {
-  return (
-    <div
-      role="presentation"
-      className="mt-3 flex items-center gap-2"
-      onClick={(event) => event.stopPropagation()}
-      onKeyDown={(event) => event.stopPropagation()}
-    >
-      <span className="text-[11px] font-medium text-gray-10 mr-1">
-        {props.opt.behaviorTitle}:
-      </span>
-      <div className="flex flex-wrap items-center gap-3">
-        {(props.opt.behaviorOptions ?? []).map((option) => (
-          <button
-            key={option.value ?? "default"}
-            type="button"
-            className={[
-              "text-[11px] transition-colors",
-              props.opt.behaviorValue === option.value
-                ? "text-gray-12 font-semibold"
-                : "text-gray-10 hover:text-gray-12",
-            ].join(" ")}
-            onClick={(event) => {
-              event.preventDefault();
-              event.stopPropagation();
-              props.onBehaviorChange(
-                { providerID: props.opt.providerID, modelID: props.opt.modelID },
-                option.value,
-              );
-            }}
-          >
-            {option.label}
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function ProviderLinkRow(props: {
-  provider: ProviderLinkItem;
-  activeIndex: number;
-  registerOptionRef: (index: number) => (el: HTMLDivElement | null) => void;
-  onSetActiveIndex: (index: number) => void;
-  onOpenSettings: () => void;
-  onClose: (options?: { restorePromptFocus?: boolean }) => void;
-}) {
-  const isKeyboardActive = props.provider.index === props.activeIndex;
-  const openProviderSettings = () => {
-    props.onClose({ restorePromptFocus: false });
-    props.onOpenSettings();
-  };
-
-  return (
-    <div
-      role="button"
-      tabIndex={0}
-      ref={props.registerOptionRef(props.provider.index)}
-      className={[
-        "group w-full text-left rounded-xl px-3 py-2.5 transition-colors cursor-pointer",
-        isKeyboardActive ? "bg-gray-2 text-gray-12" : "text-gray-10 hover:bg-gray-1/70 hover:text-gray-11",
-      ].join(" ")}
-      onMouseEnter={() => props.onSetActiveIndex(props.provider.index)}
-      onClick={openProviderSettings}
-      onKeyDown={(event: ReactKeyboardEvent<HTMLDivElement>) => {
-        if (event.key !== "Enter" && event.key !== " ") return;
-        if (event.nativeEvent.isComposing) return;
-        event.preventDefault();
-        openProviderSettings();
-      }}
-    >
-      <div className="flex items-start gap-3">
-        <ProviderIcon
-          providerId={props.provider.providerID}
-          size={16}
-          className={[
-            "mt-[1px] shrink-0 transition-colors",
-            isKeyboardActive ? "text-gray-12" : "text-gray-10 group-hover:text-gray-11",
-          ].join(" ")}
-        />
-        <div className="flex-1 min-w-0">
-          <div className="text-[13px] flex items-center justify-between gap-2 text-current">
-            <span className="truncate">{props.provider.title}</span>
-          </div>
-          <div
-            className={[
-              "mt-0.5 flex items-center gap-3 text-[11px]",
-              isKeyboardActive ? "text-gray-10" : "text-gray-9 group-hover:text-gray-10",
-            ].join(" ")}
-          >
-            <span className="truncate">{t("model_picker.connect_provider_hint")}</span>
-            <span className="ml-auto opacity-70">
-              {t("model_picker.model_count", { count: props.provider.matchCount })}
-            </span>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Small helper so downstream callers can keep the check/circle icons
-// colocated with the picker (future use for selected-state ornaments).
-export function ModelPickerSelectedIcon({ active }: { active: boolean }) {
-  return active ? <CheckCircle2 size={14} /> : <Circle size={14} />;
 }
