+++ apps/app/src/react-app/domains/settings/pages/mcp-view.tsx
@@ -39,6 +39,12 @@ import { Button } from "../../../design-system/button";
 import { ConfirmModal } from "../../../design-system/modals/confirm-modal";
 import { AddMcpModal } from "../../connections/modals/add-mcp-modal";
 import { ChromeConnectionSetupModal } from "../../connections/modals/chrome-connection-setup-modal";
+import {
+  initialMcpViewLocalState,
+  mcpViewLocalReducer,
+  type ConfigScope,
+  type McpViewLocalState,
+} from "./mcp-view-state";
 
 export type ReactMcpStatus =
   | "connected"
@@ -154,71 +160,6 @@ const serviceIconBg = (name: string) => {
   }
   return "bg-dls-hover border-dls-border";
 };
-
-type ConfigScope = "project" | "global";
-
-type McpViewLocalState = {
-  logoutOpen: boolean;
-  logoutTarget: string | null;
-  logoutBusy: boolean;
-  removeOpen: boolean;
-  removeTarget: string | null;
-  configScope: ConfigScope;
-  projectConfig: OpencodeConfigFile | null;
-  globalConfig: OpencodeConfigFile | null;
-  configError: string | null;
-  revealBusy: boolean;
-  showAdvanced: boolean;
-  addMcpModalOpen: boolean;
-  togglingMcp: string | null;
-  chromeSetupOpen: boolean;
-};
-
-type McpViewLocalAction<K extends keyof McpViewLocalState = keyof McpViewLocalState> =
-  | { type: "set"; key: K; value: SetStateAction<any> }
-  | { type: "configUnavailable" }
-  | { type: "configLoaded"; project: OpencodeConfigFile | null; global: OpencodeConfigFile | null }
-  | { type: "configLoadError"; error: string };
-
-const initialMcpViewLocalState: McpViewLocalState = {
-  logoutOpen: false,
-  logoutTarget: null,
-  logoutBusy: false,
-  removeOpen: false,
-  removeTarget: null,
-  configScope: "project",
-  projectConfig: null,
-  globalConfig: null,
-  configError: null,
-  revealBusy: false,
-  showAdvanced: false,
-  addMcpModalOpen: false,
-  togglingMcp: null,
-  chromeSetupOpen: false,
-};
-
-function mcpViewLocalReducer(
-  state: McpViewLocalState,
-  action: McpViewLocalAction,
-): McpViewLocalState {
-  switch (action.type) {
-    case "set": {
-      const current = state[action.key];
-      const next =
-        typeof action.value === "function"
-          ? (action.value as (value: typeof current) => typeof current)(current)
-          : action.value;
-      if (Object.is(current, next)) return state;
-      return { ...state, [action.key]: next };
-    }
-    case "configUnavailable":
-      return { ...state, projectConfig: null, globalConfig: null, configError: null };
-    case "configLoaded":
-      return { ...state, projectConfig: action.project, globalConfig: action.global };
-    case "configLoadError":
-      return { ...state, projectConfig: null, globalConfig: null, configError: action.error };
-  }
-}
 
 export function McpView(props: McpViewProps) {
   const showHeader = props.showHeader !== false;
