+++ apps/app/src/react-app/domains/workspace/create-workspace-modal.tsx
@@ -22,6 +22,11 @@ import {
 import type { WorkspacePreset } from "../../../app/types";
 import { usePlatform } from "../../kernel/platform";
 import { CreateWorkspaceLocalPanel } from "./create-workspace-local-panel";
+import {
+  createInitialWorkspaceLocalState,
+  createWorkspaceLocalReducer,
+  type CreateWorkspaceLocalState,
+} from "./create-workspace-modal-state";
 import { CreateWorkspaceSharedPanel } from "./create-workspace-shared-panel";
 import {
   modalBodyClass,
@@ -87,70 +92,6 @@ function workerSecondaryLine(worker: DenWorkerSummary) {
   const parts = [worker.provider?.trim() || t("dashboard.cloud_worker")];
   if (worker.instanceUrl?.trim()) parts.push(worker.instanceUrl.trim());
   return parts.join(" · ");
-}
-
-type CreateWorkspaceLocalState = {
-  screen: CreateWorkspaceScreen;
-  selectedFolder: string | null;
-  pickingFolder: boolean;
-  showProgressDetails: boolean;
-  now: number;
-  cloudSettings: ReturnType<typeof readDenSettings>;
-  remoteUrl: string;
-  remoteToken: string;
-  remoteDisplayName: string;
-  remoteTokenVisible: boolean;
-  orgs: DenOrgSummary[];
-  activeOrgId: string;
-  orgsBusy: boolean;
-  orgsError: string | null;
-  workers: DenWorkerSummary[];
-  workersBusy: boolean;
-  workersError: string | null;
-  openingWorkerId: string | null;
-  workerSearch: string;
-};
-
-type CreateWorkspaceLocalAction<K extends keyof CreateWorkspaceLocalState = keyof CreateWorkspaceLocalState> =
-  | { type: "set"; key: K; value: SetStateAction<CreateWorkspaceLocalState[K]> }
-  | { type: "reset"; settings: ReturnType<typeof readDenSettings> };
-
-function createInitialWorkspaceLocalState(settings = readDenSettings()): CreateWorkspaceLocalState {
-  return {
-    screen: "chooser",
-    selectedFolder: null,
-    pickingFolder: false,
-    showProgressDetails: false,
-    now: Date.now(),
-    cloudSettings: settings,
-    remoteUrl: "",
-    remoteToken: "",
-    remoteDisplayName: "",
-    remoteTokenVisible: false,
-    orgs: [],
-    activeOrgId: settings.activeOrgId?.trim() ?? "",
-    orgsBusy: false,
-    orgsError: null,
-    workers: [],
-    workersBusy: false,
-    workersError: null,
-    openingWorkerId: null,
-    workerSearch: "",
-  };
-}
-
-function createWorkspaceLocalReducer(
-  state: CreateWorkspaceLocalState,
-  action: CreateWorkspaceLocalAction,
-): CreateWorkspaceLocalState {
-  if (action.type === "reset") return createInitialWorkspaceLocalState(action.settings);
-  const current = state[action.key];
-  const next =
-    typeof action.value === "function"
-      ? (action.value as (value: typeof current) => typeof current)(current)
-      : action.value;
-  if (Object.is(current, next)) return state;
-  return { ...state, [action.key]: next };
 }
 
 export function CreateWorkspaceModal(props: CreateWorkspaceModalProps) {
