+++ apps/app/src/react-app/kernel/server-provider.tsx
@@ -14,6 +14,7 @@ import { createOpencodeClient } from "@opencode-ai/sdk/v2/client";
 import { desktopFetch } from "../../app/lib/desktop";
 import { isWebDeployment } from "../../app/lib/openwork-deployment";
 import { isDesktopRuntime } from "../../app/utils";
+import { initialServerState, serverReducer } from "./server-provider-state";
 
 export function normalizeServerUrl(input: string): string | undefined {
   const trimmed = input.trim();
@@ -90,50 +91,6 @@ type ServerProviderProps = {
   children: ReactNode;
   defaultUrl: string;
 };
-
-type ServerState = {
-  list: string[];
-  active: string;
-  healthy: boolean | undefined;
-};
-
-type ServerAction =
-  | { type: "ready"; list: string[]; active: string }
-  | { type: "active"; active: string }
-  | { type: "add"; url: string }
-  | { type: "remove"; url: string }
-  | { type: "healthy"; healthy: boolean | undefined };
-
-const initialServerState: ServerState = {
-  list: [],
-  active: "",
-  healthy: undefined,
-};
-
-function serverReducer(state: ServerState, action: ServerAction): ServerState {
-  switch (action.type) {
-    case "ready":
-      return { ...state, list: action.list, active: action.active };
-    case "active":
-      return { ...state, active: action.active };
-    case "add":
-      return {
-        ...state,
-        list: state.list.includes(action.url) ? state.list : [...state.list, action.url],
-        active: action.url,
-      };
-    case "remove": {
-      const list = state.list.filter((item) => item !== action.url);
-      return {
-        ...state,
-        list,
-        active: state.active === action.url ? list[0] ?? "" : state.active,
-      };
-    }
-    case "healthy":
-      return { ...state, healthy: action.healthy };
-  }
-}
 
 export function ServerProvider({ children, defaultUrl }: ServerProviderProps) {
   const [{ list, active, healthy }, dispatchServer] = useReducer(serverReducer, initialServerState);
